# Task 2/prtfoliomanager/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Portfolio Manager API",
    "description": "Complete API for portfolio management with projects, skills, categories, search, and dashboard stats.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/projects/search": {
      "get": {
        "summary": "Search projects by name, technology, or category",
        "parameters": [
          { "name": "q", "in": "query", "description": "Search by name/description" },
          { "name": "tech", "in": "query", "description": "Filter by technology" },
          { "name": "category", "in": "query", "description": "Filter by category name" }
        ]
      }
    },
...
  }
}